优化剩余接口

+++ Common/Manager.Core/Models/Blogs/BlogImage.cs
@@ -92,6 +92,6 @@ namespace Manager.Core.Models.Blogs
         /// </summary>
         [NotMapped]
         [JsonProperty("like")]
-        public int? Like { get; set; }
+        public long? Like { get; set; }
     }
 }
